dashboard: refactor collection-view into table-view
