+++ packages/dashboard/src/admin/comps/common-fields.jsx
@@ -79,14 +79,14 @@ const readable_span_cls = 'pr-3 py-2 max-w-[18rem] \
 overflow-x-auto inline-block whitespace-nowrap';
 
 /**
- * This is used in `CollectionView`
+ * This is used in `TableSchemaView`
  * 
  * @typedef {object} InternalSpanParams
  * @prop {string} [className]
  * @prop {string} [extra]
  * @prop {React.ReactNode} [children]
  * 
- * @typedef {import('./collection-view.jsx').CollectionViewComponentParams<string> & 
+ * @typedef {import('./table-schema-view.jsx').TableSchemaViewComponentParams<string> & 
 *   InternalSpanParams & 
 *   React.DetailedHTMLProps<React.HTMLAttributes<HTMLParagraphElement>, HTMLParagraphElement>
 * } SpanParams
@@ -110,7 +110,7 @@ export const Span = (
 }
 
 /**
- * This is used in `CollectionView`
+ * This is used in `TableSchemaView`
  * 
  * @typedef {object} InternalSpanArrayParams
  * @prop {string} [className]
@@ -118,7 +118,7 @@ export const Span = (
  * @prop {string} [delimiter]
  * @prop {(value: any) => string} [name_fn]
  * 
- * @typedef {import('./collection-view.jsx').CollectionViewComponentParams<any[]> & 
+ * @typedef {import('./table-schema-view.jsx').TableSchemaViewComponentParams<any[]> & 
 * InternalSpanArrayParams & 
 * React.DetailedHTMLProps<React.HTMLAttributes<HTMLParagraphElement>, HTMLParagraphElement>
 * } SpanArrayParams
@@ -153,8 +153,8 @@ export const SpanArray = (
 
 /**
  * 
- * This component is used in a `CollectionView` 
- * @param {import('./collection-view.jsx').CollectionViewComponentParams<string>
+ * This component is used in a `TableSchemaView` 
+ * @param {import('./table-schema-view.jsx').TableSchemaViewComponentParams<string>
 * } params 
 */
 export const TimeStampView = ({field, value, ...rest}) => {
@@ -444,8 +444,8 @@ export const withCard = (
 
 
 /**
- * This component is used in a `CollectionView` 
- * @param {import('./collection-view.jsx').CollectionViewComponentParams
+ * This component is used in a `TableSchemaView` 
+ * @param {import('./table-schema-view.jsx').TableSchemaViewComponentParams
  * } params 
  */
 export const RecordActions = (
