+++ packages/dashboard/src/admin/comps/resource-browse.jsx
