+++ packages/dashboard/src/admin/pages/posts.jsx
@@ -1,9 +1,9 @@
-import CollectionView from '@/admin/comps/collection-view.jsx'
 import ShowIf from '@/admin/comps/show-if.jsx'
 import { BottomActions, TopActions } from '@/admin/comps/collection-actions.jsx'
 import { Span, RecordActions, TimeStampView } from '@/admin/comps/common-fields.jsx'
 import { Title } from '@/admin/comps/common-ui.jsx'
 import useCollectionsActions from '@/admin/hooks/useCollectionsActions.js'
+import { TableSchemaView } from '../comps/table-schema-view.jsx'
 
 const schema_fields = [
   { 
@@ -44,7 +44,7 @@ export default ({}) => {
             createLink='/pages/posts/create'
             searchTitle='Search by name, values...' 
             isLoading={loading} />
-        <CollectionView 
+        <TableSchemaView 
             context={context} 
             data={page} 
             fields={schema_fields} />
