+++ packages/dashboard/src/admin/pages/products.jsx
@@ -1,4 +1,4 @@
-import CollectionView from '@/admin/comps/collection-view.jsx'
+import { TableSchemaView } from '@/admin/comps/table-schema-view.jsx'
 import ShowIf from '@/admin/comps/show-if.jsx'
 import { BottomActions, TopActions } from '@/admin/comps/collection-actions.jsx'
 import { RecordActions, Span, SpanArray, 
@@ -83,7 +83,7 @@ export default ({}) => {
             createLink='/pages/products/create'
             searchTitle='Search by Name, Handle, Tag values, Collections...' 
             isLoading={loading} />
-        <CollectionView 
+        <TableSchemaView 
             context={context} 
             data={page} 
             fields={schema_fields} />
